requirements.txt
werkzeug==0.8.3
flask==0.9
Flask-Login
Flask-Mail
Flask-WTF
Flask-OAuth
Flask-Bcrypt
Flask-DebugToolbar
Flask-Admin
Flask-SQLAlchemy
WTAlchemy==0.1b3
SQLAlchemy-migrate
itsdangerous
